requirements.txt
tabulate >= 0.8.2, < 0.8.999
PyYAML >= 3.12, < 3.999
